fix: pr review items
