--- CodeEdit/Features/Settings/Pages/SearchSettings/SearchSettingsIgnoreGlobPatternAddView.swift
@@ -1,51 +1,0 @@
-//
-//  SearchSettingsIgnoreGlobPatternAddView.swift
-//  CodeEdit
-//
-//  Created by Esteban on 12/10/23.
-//
-
-import SwiftUI
-
-struct SearchSettingsIgnoreGlobPatternAddView: View {
-    @Environment(\.dismiss)
-    var dismiss
-
-    @State private var globPattern: String = ""
-    @FocusState private var globPatternIsFocused: Bool
-
-    var body: some View {
-        VStack(spacing: 0) {
-            Form {
-                Section {
-                    VStack(alignment: .leading, spacing: 0) {
-                        TextField(
-                                "Glob Pattern",
-                                text: $globPattern
-                            )
-                            .focused($globPatternIsFocused)
-                            .onSubmit {
-                                print($globPattern)
-                            }
-                    }
-                } footer: {
-                    HStack {
-                        Spacer()
-                        Button {
-                            dismiss()
-                        } label: {
-                            Text("Cancel")
-                                .padding(.horizontal)
-                        }
-                        .buttonStyle(.borderedProminent)
-                        .controlSize(.large)
-                    }
-                    .padding(.top, 10)
-                }
-            }
-            .formStyle(.grouped)
-            .scrollDisabled(true)
-        }
-        .frame(width: 300)
-    }
-}
