+++ CodeEdit/Features/Settings/Pages/SearchSettings/SearchSettingsView.swift
@@ -30,30 +30,6 @@ struct ExcludedGlobPatternList: View {
 
     @State private var selection: GlobPattern?
 
-    func addPattern() {
-        searchSettingsModel.ignoreGlobPatterns.append(GlobPattern(value: ""))
-    }
-
-    func removePattern(_ pattern: GlobPattern?) {
-        let selectedIndex = searchSettingsModel.ignoreGlobPatterns.firstIndex {
-            $0 == selection
-        }
-
-        let removeIndex = searchSettingsModel.ignoreGlobPatterns.firstIndex {
-            $0 == selection
-        }
-
-        searchSettingsModel.ignoreGlobPatterns.removeAll {
-            pattern == $0
-        }
-
-        if selectedIndex == removeIndex && !searchSettingsModel.ignoreGlobPatterns.isEmpty && selectedIndex != nil {
-            selection = searchSettingsModel.ignoreGlobPatterns[
-                selectedIndex == 0 ? 0 : (selectedIndex ?? 1) - 1
-            ]
-        }
-    }
-
     var body: some View {
         List(selection: $selection) {
             ForEach(
@@ -130,61 +106,28 @@ struct ExcludedGlobPatternList: View {
             removePattern(selection)
         }
     }
-}
 
-struct IgnorePatternListItem: View {
-    @Binding var pattern: GlobPattern
-    @Binding var selectedPattern: GlobPattern?
-    var addPattern: () -> Void
-    var removePattern: (GlobPattern) -> Void
-    var focusedField: FocusState<String?>.Binding
-    var isLast: Bool
-
-    @State var value: String
-
-    @FocusState private var isFocused: Bool
-
-    init(
-        pattern: Binding<GlobPattern>,
-        selectedPattern: Binding<GlobPattern?>,
-        addPattern: @escaping () -> Void,
-        removePattern: @escaping (GlobPattern) -> Void,
-        focusedField: FocusState<String?>.Binding,
-        isLast: Bool
-    ) {
-        self._pattern = pattern
-        self._selectedPattern = selectedPattern
-        self.addPattern = addPattern
-        self.removePattern = removePattern
-        self.focusedField = focusedField
-        self.isLast = isLast
-        self._value = State(initialValue: pattern.wrappedValue.value)
+    func addPattern() {
+        searchSettingsModel.ignoreGlobPatterns.append(GlobPattern(value: ""))
     }
 
-    var body: some View {
-        TextField("", text: $value)
-            .focused(focusedField, equals: pattern.id.uuidString)
-            .focused($isFocused)
-            .disableAutocorrection(true)
-            .autocorrectionDisabled()
-            .labelsHidden()
-            .onSubmit {
-                if !value.isEmpty && isLast {
-                    addPattern()
-                }
-            }
-            .onChange(of: isFocused) { newIsFocused in
-                if newIsFocused {
-                    if selectedPattern != pattern {
-                        selectedPattern = pattern
-                    }
-                } else {
-                    if value.isEmpty {
-                        removePattern(pattern)
-                    } else {
-                        pattern.value = value
-                    }
-                }
-            }
+    func removePattern(_ pattern: GlobPattern?) {
+        let selectedIndex = searchSettingsModel.ignoreGlobPatterns.firstIndex {
+            $0 == selection
+        }
+
+        let removeIndex = searchSettingsModel.ignoreGlobPatterns.firstIndex {
+            $0 == selection
+        }
+
+        searchSettingsModel.ignoreGlobPatterns.removeAll {
+            pattern == $0
+        }
+
+        if selectedIndex == removeIndex && !searchSettingsModel.ignoreGlobPatterns.isEmpty && selectedIndex != nil {
+            selection = searchSettingsModel.ignoreGlobPatterns[
+                selectedIndex == 0 ? 0 : (selectedIndex ?? 1) - 1
+            ]
+        }
     }
 }
